fix: 프록시 문제로 비동기 안되는 문제 해결
