+++ src/main/java/com/tradin/module/strategy/service/StrategyService.java
@@ -13,12 +13,12 @@ import com.tradin.module.strategy.domain.repository.StrategyRepository;
 import com.tradin.module.strategy.domain.repository.dao.StrategyInfoDao;
 import com.tradin.module.strategy.service.dto.UnSubscribeStrategyDto;
 import com.tradin.module.strategy.service.dto.WebHookDto;
+import com.tradin.module.trade.service.TradeService;
 import com.tradin.module.users.domain.Users;
 import com.tradin.module.users.service.UsersService;
 import com.tradin.module.users.service.dto.SubscribeStrategyDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -33,32 +33,18 @@ public class StrategyService {
     private final HistoryService historyService;
     private final BinanceFeignService binanceFeignService;
     private final UsersService userService;
+    private final TradeService tradeService;
     private final StrategyRepository strategyRepository;
     private final AESUtils aesUtils;
 
     public void handleWebHook(WebHookDto request) {
         Strategy strategy = findByName(request.getName());
 
-        autoTrading(strategy).thenRun(() -> {
+        autoTrading(strategy.getName(), strategy.getCurrentPosition().getTradingType()).thenRun(() -> {
             closeOngoingHistory(strategy, request.getPosition());
             createNewHistory(strategy, request.getPosition());
             updateStrategyMetaData(strategy, request.getPosition());
         });
-    }
-
-    @Async
-    public CompletableFuture<Void> autoTrading(Strategy strategy) {
-        List<Users> autoTradingSubscribers = userService.findAutoTradingSubscriberByStrategyName(
-                strategy.getName());
-
-        for (Users autoTradingSubscriber : autoTradingSubscribers) {
-            String apiKey = getDecryptedKey(autoTradingSubscriber.getBinanceApiKey());
-            String secretKey = getDecryptedKey(autoTradingSubscriber.getBinanceSecretKey());
-
-            trade(autoTradingSubscriber, strategy, apiKey, secretKey);
-        }
-
-        return CompletableFuture.completedFuture(null);
     }
 
     public FindStrategiesInfoResponseDto findFutureStrategiesInfo() {
@@ -92,6 +78,11 @@ public class StrategyService {
 
     }
 
+    private CompletableFuture<Void> autoTrading(String name, TradingType tradingType) {
+        tradeService.autoTrading(name, tradingType);
+        return null;
+    }
+
     private static String getSideFromUserCurrentPosition(Users savedUser) {
         return savedUser.getCurrentPositionType().equals(TradingType.LONG) ? "SELL" : "BUY";
     }
@@ -104,103 +95,12 @@ public class StrategyService {
 //        strategyRepository.save(strategy);
 //    }
 
-    private void trade(Users user, Strategy strategy, String apiKey, String secretKey) {
-        TradingType currentPosition = getCurrentPosition(strategy);
-
-        if (currentPosition == TradingType.LONG) {
-            processLongPosition(apiKey, secretKey, "SELL", user);
-        } else if (currentPosition == TradingType.SHORT) {
-            processShortPosition(apiKey, secretKey, "BUY", user);
-        }
-
-    }
-
-    private TradingType getCurrentPosition(Strategy strategy) {
-        return strategy.getCurrentPosition().getTradingType();
-    }
-
-    private void processLongPosition(String apikey, String secretKey, String side, Users user) {
-        if (isUserTradingTypeContainsShort(user)) {
-            int orderQuantity = calculateOrderQuantity(apikey, secretKey, user.getLeverage(), user.getQuantityRate());
-
-            if (isUserPositionExist(user.getCurrentPositionType())) {
-                switchAndChangeCurrentPosition(apikey, secretKey, side, orderQuantity, user, TradingType.SHORT);
-            } else {
-                openAndChangeCurrentPosition(apikey, secretKey, side, orderQuantity, user, TradingType.SHORT);
-            }
-        } else {
-            closeAndChangeCurrentPosition(apikey, secretKey, side, user, TradingType.NONE);
-        }
-    }
-
-    private void processShortPosition(String apikey, String secretKey, String side, Users user) {
-        if (isUserTradingTypeContainsLong(user)) {
-            int orderQuantity = calculateOrderQuantity(apikey, secretKey, user.getLeverage(), user.getQuantityRate());
-
-            if (isUserPositionExist(user.getCurrentPositionType())) {
-                switchAndChangeCurrentPosition(apikey, secretKey, side, orderQuantity, user, TradingType.LONG);
-            } else {
-                openAndChangeCurrentPosition(apikey, secretKey, side, orderQuantity, user, TradingType.LONG);
-            }
-        } else {
-            closeAndChangeCurrentPosition(apikey, secretKey, side, user, TradingType.NONE);
-        }
-    }
-
-    private void switchAndChangeCurrentPosition(String apikey, String secretKey, String side, int orderQuantity, Users user, TradingType type) {
-        switchPosition(apikey, secretKey, side, orderQuantity);
-        changeCurrentPosition(user, type);
-    }
-
-    private void closeAndChangeCurrentPosition(String apikey, String secretKey, String side, Users user, TradingType type) {
-        closePosition(apikey, secretKey, side);
-        changeCurrentPosition(user, type);
-    }
-
-    private void openAndChangeCurrentPosition(String apikey, String secretKey, String side, int orderQuantity, Users user, TradingType type) {
-        openPosition(apikey, secretKey, side, orderQuantity);
-        changeCurrentPosition(user, type);
-    }
-
-    private void switchPosition(String apiKey, String secretKey, String side, int orderQuantity) {
-        binanceFeignService.openPosition(apiKey, secretKey, side, orderQuantity);
-    }
-
     private void closePosition(String apiKey, String secretKey, String side) {
         binanceFeignService.closePosition(apiKey, secretKey, side);
     }
 
-    private void openPosition(String apiKey, String secretKey, String side, int orderQuantity) {
-        binanceFeignService.openPosition(apiKey, secretKey, side, orderQuantity);
-    }
-
-    private static void changeCurrentPosition(Users user, TradingType tradingType) {
-        user.changeCurrentPosition(tradingType);
-    }
-
-    private int calculateOrderQuantity(String apiKey, String secretKey, int leverage, int quantityRate) {
-        int futureAccountBalance = binanceFeignService.getFutureAccountBalance(apiKey, secretKey);
-        return futureAccountBalance * leverage * quantityRate / 100;
-    }
-
-    private boolean isUserPositionExist(TradingType tradingType) {
-        return tradingType != TradingType.NONE;
-    }
-
-    private boolean isUserTradingTypeContainsLong(Users user) {
-        return user.getTradingType() == TradingType.LONG || user.getTradingType() == TradingType.BOTH;
-    }
-
-    private boolean isUserTradingTypeContainsShort(Users user) {
-        return user.getTradingType() == TradingType.SHORT || user.getTradingType() == TradingType.BOTH;
-    }
-
     private String getEncryptedKey(String key) {
         return aesUtils.encrypt(key);
-    }
-
-    private String getDecryptedKey(String encryptedKey) {
-        return aesUtils.decrypt(encryptedKey);
     }
 
     private Users getUserFromSecurityContext() {
@@ -238,4 +138,9 @@ public class StrategyService {
         return strategyRepository.findSpotStrategiesInfoDao()
                 .orElseThrow(() -> new TradinException(ExceptionMessage.NOT_FOUND_ANY_STRATEGY_EXCEPTION));
     }
+
+    private boolean isUserPositionExist(TradingType tradingType) {
+        return tradingType != TradingType.NONE;
+    }
+
 }
